style: 优化毛玻璃效果与全局UI组件风格

+++ lib/controllers/music_library_controller.dart
@@ -469,6 +469,9 @@ class MusicLibraryController extends ChangeNotifier {
   }
 
   Future<void> deletePlaylist(MusicPlaylist playlist) async {
+    if (playlist.isDefault) {
+      throw const KugouApiException('默认收藏歌单不能删除');
+    }
     await repository.deletePlaylist(playlist);
     playlists = playlists
         .where((item) => item.listId != playlist.listId)
